+++ package/butterfly/lib/butterfly.dart
@@ -22,12 +22,10 @@ import 'package:meta/meta.dart';
 import 'src/framework.dart';
 import 'src/key.dart';
 import 'src/surface.dart';
-import 'src/widgets.dart';
 
 export 'src/framework.dart';
 export 'src/key.dart';
 export 'src/surface.dart';
-export 'src/widgets.dart';
 
 part 'src/tree.dart';
 part 'src/util.dart';
@@ -35,6 +33,6 @@ part 'src/basic/container.dart';
 part 'src/basic/layout.dart';
 part 'src/basic/common.dart';
 
-void runApp(Node widget, Surface host) {
+void runApp(Widget widget, Surface host) {
   new Tree(widget, host).renderFrame();
 }
